Refactor code as per the PR
