+++ modules/siddhi-core/src/main/java/org/wso2/siddhi/core/stream/StreamJunction.java
@@ -33,7 +33,7 @@ import org.wso2.siddhi.core.stream.input.InputProcessor;
 import org.wso2.siddhi.core.stream.output.StreamCallback;
 import org.wso2.siddhi.core.util.SiddhiConstants;
 import org.wso2.siddhi.core.util.statistics.ThroughputTracker;
-import org.wso2.siddhi.core.util.timestamp.EventTimeMillisTimestampGenerator;
+import org.wso2.siddhi.core.util.timestamp.EventBasedTimeMillisTimestampGenerator;
 import org.wso2.siddhi.query.api.annotation.Annotation;
 import org.wso2.siddhi.query.api.definition.StreamDefinition;
 import org.wso2.siddhi.query.api.exception.DuplicateAnnotationException;
@@ -199,7 +199,7 @@ public class StreamJunction {
     private void sendData(long timeStamp, Object[] data) {
         // Set timestamp to system if Siddhi is in playback mode
         if (executionPlanContext.isPlayback()) {
-            ((EventTimeMillisTimestampGenerator) this.executionPlanContext.getTimestampGenerator()).setTimestamp(timeStamp);
+            ((EventBasedTimeMillisTimestampGenerator) this.executionPlanContext.getTimestampGenerator()).setCurrentTimestamp(timeStamp);
         }
         if (throughputTracker != null) {
             throughputTracker.eventIn();
