+++ modules/siddhi-core/src/main/java/org/wso2/siddhi/core/util/Scheduler.java
@@ -29,7 +29,7 @@ import org.wso2.siddhi.core.query.input.stream.single.EntryValveProcessor;
 import org.wso2.siddhi.core.util.lock.LockWrapper;
 import org.wso2.siddhi.core.util.snapshot.Snapshotable;
 import org.wso2.siddhi.core.util.statistics.LatencyTracker;
-import org.wso2.siddhi.core.util.timestamp.EventTimeMillisTimestampGenerator;
+import org.wso2.siddhi.core.util.timestamp.EventBasedTimeMillisTimestampGenerator;
 
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
@@ -59,7 +59,7 @@ public class Scheduler implements Snapshotable {
         this.executionPlanContext = executionPlanContext;
 
         if (this.executionPlanContext.isPlayback()) {
-            ((EventTimeMillisTimestampGenerator) this.executionPlanContext.getTimestampGenerator()).addTimeChangeListener(new EventTimeMillisTimestampGenerator.TimeChangeListener() {
+            ((EventBasedTimeMillisTimestampGenerator) this.executionPlanContext.getTimestampGenerator()).addTimeChangeListener(new EventBasedTimeMillisTimestampGenerator.TimeChangeListener() {
                 @Override
                 public void onTimeChange(long currentTimestamp) {
                     Long lastTime = toNotifyQueue.peek();
@@ -77,16 +77,17 @@ public class Scheduler implements Snapshotable {
     public void notifyAt(long time) {
         try {
             toNotifyQueue.put(time);
-            if (!running && toNotifyQueue.size() == 1) {
+            if (!running && toNotifyQueue.size() == 1 && !this.executionPlanContext.isPlayback()) {
                 synchronized (toNotifyQueue) {
                     if (!running) {
                         running = true;
                         long timeDiff = time - executionPlanContext.getTimestampGenerator().currentTime();
                         if (timeDiff > 0) {
-                            if (!this.executionPlanContext.isPlayback()) {
-                                scheduledExecutorService.schedule(eventCaller, timeDiff, TimeUnit.MILLISECONDS);
-                            }
+//                            if (!this.executionPlanContext.isPlayback()) {
+                            scheduledExecutorService.schedule(eventCaller, timeDiff, TimeUnit.MILLISECONDS);
+//                            }
                         } else {
+                            // TODO: 11/21/16 check this
                             scheduledExecutorService.schedule(eventCaller, 0, TimeUnit.MILLISECONDS);
                         }
                     }
