+++ modules/siddhi-core/src/test/java/org/wso2/siddhi/core/managment/PlaybackTestCase.java
@@ -1,0 +1,541 @@
+/*
+ * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.siddhi.core.managment;
+
+import junit.framework.Assert;
+import org.apache.log4j.Logger;
+import org.junit.Before;
+import org.junit.Test;
+import org.wso2.siddhi.core.ExecutionPlanRuntime;
+import org.wso2.siddhi.core.SiddhiManager;
+import org.wso2.siddhi.core.event.Event;
+import org.wso2.siddhi.core.query.output.callback.QueryCallback;
+import org.wso2.siddhi.core.stream.input.InputHandler;
+import org.wso2.siddhi.core.util.EventPrinter;
+import org.wso2.siddhi.query.compiler.exception.SiddhiParserException;
+
+public class PlaybackTestCase {
+    private static final Logger log = Logger.getLogger(PlaybackTestCase.class);
+    private int inEventCount;
+    private int removeEventCount;
+    private boolean eventArrived;
+
+    @Before
+    public void init() {
+        inEventCount = 0;
+        removeEventCount = 0;
+        eventArrived = false;
+    }
+
+    @Test
+    public void playbackTest1() throws InterruptedException {
+        log.info("Playback Test 1: Playback with heartbeat disabled in query containing regular time batch window");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String cseEventStream = "" +
+                "@Plan:playback " +
+                "define stream cseEventStream (symbol string, price float, volume int);";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from cseEventStream#window.timeBatch(1 sec) " +
+                "select * " +
+                "insert all events into outputStream ;";
+
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
+
+        executionPlanRuntime.addCallback("query1", new QueryCallback() {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+                EventPrinter.print(timeStamp, inEvents, removeEvents);
+                if (inEventCount == 0) {
+                    Assert.assertTrue("Remove Events will only arrive after the second time period. ", removeEvents == null);
+                }
+                if (inEvents != null) {
+                    inEventCount = inEventCount + inEvents.length;
+                }
+                if (removeEvents != null) {
+                    removeEventCount = removeEventCount + removeEvents.length;
+                }
+                eventArrived = true;
+            }
+
+        });
+
+        InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
+        executionPlanRuntime.start();
+        long timestamp = System.currentTimeMillis();
+        inputHandler.send(timestamp, new Object[]{"IBM", 700f, 0});
+
+        timestamp += 500;
+        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 1});
+
+        timestamp += 500;   // 1 sec passed
+        inputHandler.send(timestamp, new Object[]{"GOOGLE", 85.0f, 1});
+
+        timestamp += 1000;   // Another 1 sec passed
+        inputHandler.send(timestamp, new Object[]{"ORACLE", 90.5f, 1});
+
+        Thread.sleep(100);
+
+        Assert.assertEquals(3, inEventCount);
+        Assert.assertEquals(2, removeEventCount);
+        Assert.assertTrue(eventArrived);
+        executionPlanRuntime.shutdown();
+
+    }
+
+    @Test
+    public void playbackTest2() throws InterruptedException {
+        log.info("Playback Test 2: Playback with heartbeat disabled in query with start time enabled time batch window");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String cseEventStream = "" +
+                "@Plan:playback " +
+                "define stream cseEventStream (symbol string, price float, volume int);";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from cseEventStream#window.timeBatch(2 sec , 0) " +
+                "select symbol, sum(price) as sumPrice, volume " +
+                "insert into outputStream ;";
+
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
+
+        executionPlanRuntime.addCallback("query1", new QueryCallback() {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+                EventPrinter.print(timeStamp, inEvents, removeEvents);
+                if (inEventCount == 0) {
+                    Assert.assertTrue("Remove Events will only arrive after the second time period. ", removeEvents == null);
+                }
+                if (inEvents != null) {
+                    inEventCount = inEventCount + inEvents.length;
+                } else if (removeEvents != null) {
+                    removeEventCount = removeEventCount + removeEvents.length;
+                }
+                eventArrived = true;
+            }
+
+        });
+
+        InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
+        executionPlanRuntime.start();
+        // Start sending events in the beginning of a cycle
+        long timestamp = 0;
+        inputHandler.send(timestamp, new Object[]{"IBM", 700f, 0});
+        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 1});
+        timestamp += 8500;
+        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 1});
+        inputHandler.send(timestamp, new Object[]{"II", 60.5f, 1});
+        timestamp += 13000;
+        inputHandler.send(timestamp, new Object[]{"TT", 60.5f, 1});
+        inputHandler.send(timestamp, new Object[]{"YY", 60.5f, 1});
+        timestamp += 5000;
+        inputHandler.send(timestamp, new Object[]{"ZZ", 0.0f, 0});
+
+        Thread.sleep(100);
+
+        Assert.assertEquals(3, inEventCount);
+        Assert.assertEquals(0, removeEventCount);
+        Assert.assertTrue(eventArrived);
+        executionPlanRuntime.shutdown();
+
+    }
+
+    @Test
+    public void playbackTest3() throws InterruptedException {
+        log.info("Playback Test 3: Playback with heartbeat enabled");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String cseEventStream = "" +
+                "@Plan:playback(idleTime = '100 millisecond', increment = '2 sec') " +
+                "define stream cseEventStream (symbol string, price float, volume int);";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from cseEventStream#window.timeBatch(2 sec , 0) " +
+                "select symbol, sum(price) as sumPrice, volume " +
+                "insert into outputStream ;";
+
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
+
+        executionPlanRuntime.addCallback("query1", new QueryCallback() {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+                EventPrinter.print(timeStamp, inEvents, removeEvents);
+                if (inEventCount == 0) {
+                    Assert.assertTrue("Remove Events will only arrive after the second time period. ", removeEvents == null);
+                }
+                if (inEvents != null) {
+                    inEventCount = inEventCount + inEvents.length;
+                } else if (removeEvents != null) {
+                    removeEventCount = removeEventCount + removeEvents.length;
+                }
+                eventArrived = true;
+            }
+
+        });
+
+        InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
+        executionPlanRuntime.start();
+        // Start sending events in the beginning of a cycle
+        long timestamp = 0;
+        inputHandler.send(timestamp, new Object[]{"IBM", 700f, 0});
+        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 1});
+        timestamp += 8500;
+        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 1});
+        inputHandler.send(timestamp, new Object[]{"II", 60.5f, 1});
+        timestamp += 13000;
+        inputHandler.send(timestamp, new Object[]{"TT", 60.5f, 1});
+        inputHandler.send(timestamp, new Object[]{"YY", 60.5f, 1});
+
+        Thread.sleep(200);  // Anything more than 100 is enough. Used 200 to be on safe side
+
+        Assert.assertEquals(3, inEventCount);
+        Assert.assertEquals(0, removeEventCount);
+        Assert.assertTrue(eventArrived);
+        executionPlanRuntime.shutdown();
+
+    }
+
+    @Test
+    public void playbackTest4() throws InterruptedException {
+        log.info("Playback Test 4: Playback with query joining two windows");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        String streams = "" +
+                "@Plan:playback(idleTime = '100 millisecond', increment = '1 sec') " +
+                "define stream cseEventStream (symbol string, price float, volume int); " +
+                "define stream twitterStream (user string, tweet string, company string); ";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from cseEventStream#window.timeBatch(1 sec) join twitterStream#window.timeBatch(1 sec) " +
+                "on cseEventStream.symbol== twitterStream.company " +
+                "select cseEventStream.symbol as symbol, twitterStream.tweet, cseEventStream.price " +
+                "insert into outputStream ;";
+
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(streams + query);
+        try {
+            executionPlanRuntime.addCallback("query1", new QueryCallback() {
+                @Override
+                public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+                    EventPrinter.print(timeStamp, inEvents, removeEvents);
+                    if (inEvents != null) {
+                        inEventCount += (inEvents.length);
+                    }
+                    if (removeEvents != null) {
+                        removeEventCount += (removeEvents.length);
+                    }
+                    eventArrived = true;
+                }
+            });
+            InputHandler cseEventStreamHandler = executionPlanRuntime.getInputHandler("cseEventStream");
+            InputHandler twitterStreamHandler = executionPlanRuntime.getInputHandler("twitterStream");
+            executionPlanRuntime.start();
+            long currentTime = System.currentTimeMillis();
+            cseEventStreamHandler.send(currentTime, new Object[]{"WSO2", 55.6f, 100});
+            twitterStreamHandler.send(currentTime, new Object[]{"User1", "Hello World", "WSO2"});
+            cseEventStreamHandler.send(currentTime, new Object[]{"IBM", 75.6f, 100});
+            currentTime += 1500;
+            cseEventStreamHandler.send(currentTime, new Object[]{"WSO2", 57.6f, 100});
+
+            Thread.sleep(200);  // Anything more than 100 is enough. Used 200 to be on safe side
+
+            Assert.assertTrue("In Events can be 1 or 2 ", inEventCount == 1 || inEventCount == 2);
+            Assert.assertEquals(0, removeEventCount);
+            Assert.assertTrue(eventArrived);
+        } finally {
+            executionPlanRuntime.shutdown();
+        }
+    }
+
+    @Test
+    public void playbackTest5() throws InterruptedException {
+        log.info("Playback Test 5: Playback enabled timeLength window with no of events less than window length and time period less than window time");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String cseEventStream = "@Plan:playback define stream cseEventStream (symbol string, price float, volume int);";
+        String query = "@info(name = 'query1') from cseEventStream#window.timeLength(4 sec,10) select symbol,price," +
+                "volume insert all events into outputStream ;";
+
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
+
+        executionPlanRuntime.addCallback("query1", new QueryCallback() {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+                EventPrinter.print(timeStamp, inEvents, removeEvents);
+                if (inEvents != null) {
+                    inEventCount = inEventCount + inEvents.length;
+                }
+                if (removeEvents != null) {
+                    Assert.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
+                    removeEventCount = removeEventCount + removeEvents.length;
+                }
+                eventArrived = true;
+            }
+
+        });
+
+        InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
+        executionPlanRuntime.start();
+        long timestamp = System.currentTimeMillis();
+        inputHandler.send(timestamp, new Object[]{"IBM", 700f, 1});
+        timestamp += 500;
+        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 2});
+        timestamp += 500;
+        inputHandler.send(timestamp, new Object[]{"IBM", 700f, 3});
+        timestamp += 500;
+        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 4});
+        timestamp += 5000;
+
+        inputHandler.send(timestamp, new Object[]{"GOOGLE", 90.5f, 5});
+
+        Assert.assertEquals(5, inEventCount);
+        Assert.assertEquals(4, removeEventCount);
+        Assert.assertTrue(eventArrived);
+        executionPlanRuntime.shutdown();
+    }
+
+    @Test
+    public void playbackTest6() throws InterruptedException {
+        log.info("Playback Test 6: Playback with heartbeat enabled timeLength window with no of events less than window length and time period less than window time");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String cseEventStream = "@Plan:playback(idleTime = '100 millisecond', increment = '4 sec') define stream cseEventStream (symbol string, price float, volume int);";
+        String query = "@info(name = 'query1') from cseEventStream#window.timeLength(4 sec,10) select symbol,price," +
+                "volume insert all events into outputStream ;";
+
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
+
+        executionPlanRuntime.addCallback("query1", new QueryCallback() {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+                EventPrinter.print(timeStamp, inEvents, removeEvents);
+                if (inEvents != null) {
+                    inEventCount = inEventCount + inEvents.length;
+                }
+                if (removeEvents != null) {
+                    Assert.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
+                    removeEventCount = removeEventCount + removeEvents.length;
+                }
+                eventArrived = true;
+            }
+
+        });
+
+        InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
+        executionPlanRuntime.start();
+        long timestamp = System.currentTimeMillis();
+        inputHandler.send(timestamp, new Object[]{"IBM", 700f, 1});
+        timestamp += 500;
+        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 2});
+        timestamp += 500;
+        inputHandler.send(timestamp, new Object[]{"IBM", 700f, 3});
+        timestamp += 500;
+        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 4});
+
+        Thread.sleep(200);  // Anything more than 100 is enough. Used 200 to be on safe side
+
+        Assert.assertEquals(4, inEventCount);
+        Assert.assertEquals(4, removeEventCount);
+        Assert.assertTrue(eventArrived);
+        executionPlanRuntime.shutdown();
+    }
+
+    @Test
+    public void playbackTest7() throws InterruptedException {
+        log.info("Playback Test 7: Testing playback without heartbeat");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String cseEventStream = "" +
+                "@Plan:playback " +
+                "define stream cseEventStream (symbol string, price float, volume int);";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from cseEventStream#window.time(2 sec) " +
+                "select symbol,price,volume " +
+                "insert all events into outputStream ;";
+
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
+
+        executionPlanRuntime.addCallback("query1", new QueryCallback() {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+                EventPrinter.print(timeStamp, inEvents, removeEvents);
+                if (inEvents != null) {
+                    inEventCount = inEventCount + inEvents.length;
+                }
+                if (removeEvents != null) {
+                    Assert.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
+                    removeEventCount = removeEventCount + removeEvents.length;
+                }
+                eventArrived = true;
+            }
+
+        });
+
+        InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
+        executionPlanRuntime.start();
+        long timestamp = System.currentTimeMillis();
+        inputHandler.send(timestamp, new Object[]{"IBM", 700f, 0});
+        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 1});
+        timestamp += 2000;
+        inputHandler.send(timestamp, new Object[]{"GOOGLE", 0.0f, 1});
+        Thread.sleep(100);
+        Assert.assertEquals(3, inEventCount);
+        Assert.assertEquals(2, removeEventCount);
+        Assert.assertTrue(eventArrived);
+        executionPlanRuntime.shutdown();
+    }
+
+    @Test
+    public void playbackTest8() throws InterruptedException {
+        log.info("Playback Test 8: Testing playback with heartbeat enabled");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String cseEventStream = "" +
+                "@Plan:playback(idleTime = '100 millisecond', increment = '2 sec') " +
+                "define stream cseEventStream (symbol string, price float, volume int);";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from cseEventStream#window.time(2 sec) " +
+                "select symbol,price,volume " +
+                "insert all events into outputStream ;";
+
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
+
+        executionPlanRuntime.addCallback("query1", new QueryCallback() {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+                EventPrinter.print(timeStamp, inEvents, removeEvents);
+                if (inEvents != null) {
+                    inEventCount = inEventCount + inEvents.length;
+                }
+                if (removeEvents != null) {
+                    Assert.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
+                    removeEventCount = removeEventCount + removeEvents.length;
+                }
+                eventArrived = true;
+            }
+
+        });
+
+        InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
+        executionPlanRuntime.start();
+        long timestamp = System.currentTimeMillis();
+        inputHandler.send(timestamp, new Object[]{"IBM", 700f, 0});
+        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 1});
+
+        Thread.sleep(200);  // Anything more than 100 is enough. Used 200 to be on safe side
+
+        Assert.assertEquals(2, inEventCount);
+        Assert.assertEquals(2, removeEventCount);
+        Assert.assertTrue(eventArrived);
+        executionPlanRuntime.shutdown();
+    }
+
+    @Test(expected = SiddhiParserException.class)
+    public void playbackTest9() throws InterruptedException {
+        log.info("Playback Test 9: Testing playback with invalid increment time constant");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String cseEventStream = "" +
+                "@Plan:playback(idleTime = '100 millisecond', increment = '2') " +
+                "define stream cseEventStream (symbol string, price float, volume int);";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from cseEventStream#window.time(2 sec) " +
+                "select symbol,price,volume " +
+                "insert all events into outputStream ;";
+
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
+    }
+
+    @Test(expected = SiddhiParserException.class)
+    public void playbackTest10() throws InterruptedException {
+        log.info("Playback Test 10: Testing playback with invalid idleTime time constant");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String cseEventStream = "" +
+                "@Plan:playback(idleTime = '', increment = '2 sec') " +
+                "define stream cseEventStream (symbol string, price float, volume int);";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from cseEventStream#window.time(2 sec) " +
+                "select symbol,price,volume " +
+                "insert all events into outputStream ;";
+
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
+    }
+
+    @Test
+    public void playbackTest11() throws InterruptedException {
+        log.info("Playback Test 11: Testing playback with out of order events");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String cseEventStream = "" +
+                "@Plan:playback " +
+                "define stream cseEventStream (symbol string, price float, volume int);";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from cseEventStream#window.time(2 sec) " +
+                "select symbol,price,volume " +
+                "insert all events into outputStream ;";
+
+        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
+
+        executionPlanRuntime.addCallback("query1", new QueryCallback() {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+                EventPrinter.print(timeStamp, inEvents, removeEvents);
+                if (inEvents != null) {
+                    inEventCount = inEventCount + inEvents.length;
+                }
+                if (removeEvents != null) {
+                    Assert.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
+                    removeEventCount = removeEventCount + removeEvents.length;
+                }
+                eventArrived = true;
+            }
+
+        });
+
+        InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
+        executionPlanRuntime.start();
+        long timestamp = System.currentTimeMillis();
+        inputHandler.send(timestamp, new Object[]{"IBM", 700f, 0});
+        timestamp -= 1000;
+        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 1});
+
+        Thread.sleep(200);  // Anything more than 100 is enough. Used 200 to be on safe side
+
+        Assert.assertEquals(2, inEventCount);
+        Assert.assertEquals(2, removeEventCount);
+        Assert.assertTrue(eventArrived);
+        executionPlanRuntime.shutdown();
+    }
+}
