+++ modules/siddhi-core/src/test/java/org/wso2/siddhi/core/query/window/TimeLengthWindowTestCase.java
@@ -443,99 +443,4 @@ public class TimeLengthWindowTestCase {
         Assert.assertTrue(eventArrived);
         executionPlanRuntime.shutdown();
     }
-
-    @Test
-    public void timeLengthWindowTest11() throws InterruptedException {
-        log.info("Testing playback enabled timeLength window with no of events less than window length and time period less than window time");
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-
-        String cseEventStream = "@Plan:playback define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.timeLength(4 sec,10) select symbol,price," +
-                "volume insert all events into outputStream ;";
-
-        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
-
-        executionPlanRuntime.addCallback("query1", new QueryCallback() {
-            @Override
-            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
-                EventPrinter.print(timeStamp, inEvents, removeEvents);
-                if (inEvents != null) {
-                    inEventCount = inEventCount + inEvents.length;
-                }
-                if (removeEvents != null) {
-                    Assert.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
-                    removeEventCount = removeEventCount + removeEvents.length;
-                }
-                eventArrived = true;
-            }
-
-        });
-
-        InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
-        executionPlanRuntime.start();
-        long timestamp = System.currentTimeMillis();
-        inputHandler.send(timestamp, new Object[]{"IBM", 700f, 1});
-        timestamp += 500;
-        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 2});
-        timestamp += 500;
-        inputHandler.send(timestamp, new Object[]{"IBM", 700f, 3});
-        timestamp += 500;
-        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 4});
-        timestamp += 5000;
-
-        inputHandler.send(timestamp, new Object[]{"GOOGLE", 90.5f, 5});
-
-        Assert.assertEquals(5, inEventCount);
-        Assert.assertEquals(4, removeEventCount);
-        Assert.assertTrue(eventArrived);
-        executionPlanRuntime.shutdown();
-    }
-
-    @Test
-    public void timeLengthWindowTest12() throws InterruptedException {
-        log.info("Testing playback with heartbeat enabled timeLength window with no of events less than window length and time period less than window time");
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-
-        String cseEventStream = "@Plan:playback(idleTime = '100 millisecond', increment = '4 sec') define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "@info(name = 'query1') from cseEventStream#window.timeLength(4 sec,10) select symbol,price," +
-                "volume insert all events into outputStream ;";
-
-        ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(cseEventStream + query);
-
-        executionPlanRuntime.addCallback("query1", new QueryCallback() {
-            @Override
-            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
-                EventPrinter.print(timeStamp, inEvents, removeEvents);
-                if (inEvents != null) {
-                    inEventCount = inEventCount + inEvents.length;
-                }
-                if (removeEvents != null) {
-                    Assert.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
-                    removeEventCount = removeEventCount + removeEvents.length;
-                }
-                eventArrived = true;
-            }
-
-        });
-
-        InputHandler inputHandler = executionPlanRuntime.getInputHandler("cseEventStream");
-        executionPlanRuntime.start();
-        long timestamp = System.currentTimeMillis();
-        inputHandler.send(timestamp, new Object[]{"IBM", 700f, 1});
-        timestamp += 500;
-        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 2});
-        timestamp += 500;
-        inputHandler.send(timestamp, new Object[]{"IBM", 700f, 3});
-        timestamp += 500;
-        inputHandler.send(timestamp, new Object[]{"WSO2", 60.5f, 4});
-
-        Thread.sleep(200);  // Anything more than 100 is enough. Used 200 to be on safe side
-
-        Assert.assertEquals(4, inEventCount);
-        Assert.assertEquals(4, removeEventCount);
-        Assert.assertTrue(eventArrived);
-        executionPlanRuntime.shutdown();
-    }
 }
